requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
structlog==23.2.0
spacy==3.5.4
playwright==1.40.0
beautifulsoup4==4.12.2
pydantic==2.5.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
